refactor(core): move utxo recover method to separate file
The class method implementation was using deprecated coroutine
wrappers and was way too long.

Issue: BG-34381
